+++ sickrage/core/processors/post_processor.py
@@ -392,7 +392,7 @@ class PostProcessor(object):
                 moveFile(cur_file_path, new_file_path)
                 chmodAsParent(new_file_path)
             except (IOError, OSError) as e:
-                self._log("Unable to move file " + cur_file_path + " to " + new_file_path + ": {}".format(e),
+                self._log("Unable to move file {} to {}: {}".format(cur_file_path, new_file_path, e),
                           sickrage.app.log.ERROR)
                 raise
 
@@ -416,7 +416,7 @@ class PostProcessor(object):
                 copyFile(cur_file_path, new_file_path)
                 chmodAsParent(new_file_path)
             except (IOError, OSError) as e:
-                self._log("Unable to copy file " + cur_file_path + " to " + new_file_path + ": {}".format(e),
+                self._log("Unable to copy file {} to {}: {}".format(cur_file_path, new_file_path, e),
                           sickrage.app.log.ERROR)
                 raise
 
@@ -438,11 +438,13 @@ class PostProcessor(object):
             self._log("Hard linking file from " + cur_file_path + " to " + new_file_path,
                       sickrage.app.log.DEBUG)
             try:
+                if os.path.exists(new_file_path):
+                    os.remove(new_file_path)
+
                 hardlinkFile(cur_file_path, new_file_path)
                 chmodAsParent(new_file_path)
             except (IOError, OSError) as e:
-                self._log("Unable to link file {} to {}: {}".format(cur_file_path, new_file_path,
-                                                                    e)), sickrage.app.log.ERROR
+                self._log("Unable to hardlink file {} to {}: {}".format(cur_file_path, new_file_path, e), sickrage.app.log.ERROR)
                 raise
 
         self._combined_file_operation(file_path, new_path, new_base_name, associated_files, action=_int_hard_link,
@@ -466,7 +468,7 @@ class PostProcessor(object):
                 moveAndSymlinkFile(cur_file_path, new_file_path)
                 chmodAsParent(new_file_path)
             except (IOError, OSError) as e:
-                self._log("Unable to link file " + cur_file_path + " to " + new_file_path + ": {}".format(e),
+                self._log("Unable to move and symlink file {} to {}: {}".format(cur_file_path, new_file_path, e),
                           sickrage.app.log.ERROR)
                 raise
 
@@ -477,8 +479,8 @@ class PostProcessor(object):
         """
         symlink destination to source location, and set proper permissions
 
-        :param file_path: The full path of the media file to move
-        :param new_path: Destination path where we want to move the file to create a symbolic link to
+        :param file_path: The full path of the media file to link
+        :param new_path: Destination path where we want to create a symbolic link to
         :param new_base_name: The base filename (no extension) to use during the link. Use None to keep the same name.
         :param associated_files: Boolean, whether we should move similarly-named files too
         """
@@ -488,10 +490,13 @@ class PostProcessor(object):
             self._log("Creating then symbolic linking file from " + new_file_path + " to " + cur_file_path,
                       sickrage.app.log.DEBUG)
             try:
+                if os.path.exists(new_file_path):
+                    os.remove(new_file_path)
+
                 symlink(cur_file_path, new_file_path)
                 chmodAsParent(cur_file_path)
             except (IOError, OSError) as e:
-                self._log("Unable to link file " + cur_file_path + " to " + new_file_path + ": " + str(e),
+                self._log("Unable to symlink file {} to {}: {}".format(cur_file_path, new_file_path, e),
                           sickrage.app.log.ERROR)
                 raise